自作関数データの作成
